refactor UserDetail in a component to make profile page easier
